Add a set command.
